Changed most numerical values to float, membership dues is still accepted as an integer, but is stored as a float for late calculations. Treasury funds, proposal funding, funding_goal, membership dues are all stored as floats. This is to enable fractional membership collection.

+++ db/migrate/20190504053612_create_proposals.rb
@@ -5,8 +5,8 @@ class CreateProposals < ActiveRecord::Migration[5.2]
     create_table :proposals do |t|
       t.string :title
       t.text :description
-      t.integer :funding_goal
-      t.integer :funding, default: 0
+      t.float :funding_goal
+      t.float :funding, default: 0
       t.boolean :active, default: true
       t.date :expiration
       t.integer :user_id
