+++ db/migrate/20190505053621_create_treasuries.rb
@@ -3,7 +3,7 @@
 class CreateTreasuries < ActiveRecord::Migration[5.2]
   def change
     create_table :treasuries do |t|
-      t.integer :funds
+      t.float :funds
 
       t.timestamps
     end
